requirements.txt
python-dotenv
openai
langchain
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
langchain-openai
pandas
PyPDF2